fix compilation on i386

+++ src/test/IWASnappyStreamTest.cpp
@@ -36,7 +36,7 @@ void assertCompressed(const string &message, const unsigned char *const expected
   unsigned long uncompressedSize = 0;
   const unsigned char *const uncompressed = uncompressedStream->read(expectedSize, uncompressedSize);
   assert(uncompressed);
-  CPPUNIT_ASSERT_EQUAL_MESSAGE(message + ": size", expectedSize, uncompressedSize);
+  CPPUNIT_ASSERT_EQUAL_MESSAGE(message + ": size", expectedSize, size_t(uncompressedSize));
   CPPUNIT_ASSERT_MESSAGE(message + ": input exhausted", uncompressedStream->isEnd());
   CPPUNIT_ASSERT_MESSAGE(message + ": content", std::equal(expected, expected + expectedSize, uncompressed));
 }
@@ -48,7 +48,7 @@ void assertCompressedFull(const string &message, const unsigned char *const expe
   unsigned long uncompressedSize = 0;
   const unsigned char *const uncompressed = uncompressedStream.read(expectedSize, uncompressedSize);
   assert(uncompressed);
-  CPPUNIT_ASSERT_EQUAL_MESSAGE(message + ": size", expectedSize, uncompressedSize);
+  CPPUNIT_ASSERT_EQUAL_MESSAGE(message + ": size", expectedSize, size_t(uncompressedSize));
   CPPUNIT_ASSERT_MESSAGE(message + ": input exhausted", uncompressedStream.isEnd());
   CPPUNIT_ASSERT_MESSAGE(message + ": content", std::equal(expected, expected + expectedSize, uncompressed));
 }
